Inscripciones..Estudiantes..Pagos

+++ database/migrations/2024_11_06_153703_create_colegios_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->string('codModular');
             $table->string('modalidad');
             $table->string('gestion');
-            $table->int('latitud');
-            $table->int('longitud');
+            $table->decimal('latitud', 10, 6);
+            $table->decimal('longitud', 10, 6);
             $table->foreignId('Distrito_idDistrito')->constrained('distritos');
         });
     }
